Fix clippy warnings and apply cargo fmt
- Remove unused import (setup_screen::*)
- Use #[derive(Default)] for ProcessState and ProcessField enums
- Simplify if_same_then_else cases in process.rs and row_print.rs
- Replace manual patterns with idiomatic Rust:
  - .max().min() -> .clamp()
  - (n + d - 1) / d -> .div_ceil()
  - manual Option::map -> .map()
  - starts_with + slice -> .strip_prefix()
  - len() > 0 -> !is_empty()
  - manual range checks -> Range::contains()
  - b"\0" -> c"" literal
- Remove unnecessary casts and Box wrapper
- Use mem::take() instead of drain().collect()
- Collapse nested if statements

Remaining 8 warnings are intentional (too_many_arguments for
internal functions, enum sentinel naming).

+++ src/platform/darwin.rs
@@ -804,7 +804,7 @@ pub fn scan_processes_with_settings(machine: &mut Machine, update_process_names:
 
             // CPU time (in nanoseconds -> convert to hundredths of a second)
             let total_time_ns = task_info.pti_total_user + task_info.pti_total_system;
-            process.time = (total_time_ns / 10_000_000) as u64; // ns to centiseconds
+            process.time = total_time_ns / 10_000_000; // ns to centiseconds
 
             // Thread count
             process.nlwp = task_info.pti_threadnum as i64;
